Version 2.1
I have totally redesigned it to work seamlessly with vim's cd and cdpath features. You can now pushd into a diretory that is in the cdpath. Also added new commands to manage the cdpath itself, without needing to change the .vimrc. Removed unnecessary command-line abbreviations. Removed Cdp command as it is a duplicate of "cd -" command and renamed Cdc command as Cdp. The script now requires genutils.vim for storing and retrieving cdpath.